MainMenuTransition
This is for a main menu and not in game. It only works to change a scene to go to scene 1, but you can just change the name of the script IN the script, the number scene in the PlayGame() to whatever scene you want and you can change it to any scene.